Update config.json
Changed default for `is_licensed` from `True` to `False`

This blocks the radio from transmitting automatically when `main.py` is run. This avoids possible damage to the radio if there is no antenna attached and mitigates the risk that an illegal transmission is made if no licensed amateur radio operator is present.